Complete refactor of the code to make it more organized. Too many changes to list.
